Make ServiceSpecialityLogoLink use the Gatsby link and style
